+++ packages/smarthr-ui/src/components/Dialog/StepFormDialog/StepFormDialog.tsx
@@ -17,7 +17,7 @@ import {
   StepFormDialogContentInner,
   type StepFormDialogContentInnerProps,
 } from './StepFormDialogContentInner'
-import { StepFormDialogProvider, type StepItem } from './StepFormDialogProvider'
+import { StepFormDialogProvider } from './StepFormDialogProvider'
 
 import type { FocusTrapRef } from '../FocusTrap'
 import type { DialogProps /** コンテンツなにもないDialogの基本props */ } from '../types'
@@ -71,15 +71,12 @@ export const StepFormDialog: FC<Props> = ({
     }
   }, [isOpen, onClickClose])
 
-  const actualOnSubmitAction = useCallback(
-    (close: () => void, e: FormEvent<HTMLFormElement>, currentStep: StepItem) => {
+  const onDelegateSubmit = useCallback(
+    (e: FormEvent<HTMLFormElement>, helpers: Parameters<typeof onSubmit>[1]) => {
       if (isOpen) {
         focusTrapRef.current?.focus()
-
-        return onSubmit(close, e, currentStep)
+        onSubmit(e, helpers)
       }
-
-      return undefined
     },
     [onSubmit, isOpen],
   )
@@ -101,7 +98,6 @@ export const StepFormDialog: FC<Props> = ({
         onPressEscape={closeDisabled ? undefined : onPressEscape}
         focusTrapRef={focusTrapRef}
       >
-        {/* eslint-disable-next-line smarthr/a11y-delegate-element-has-role-presentation */}
         <StepFormDialogContentInner
           heading={heading}
           contentBgColor={contentBgColor}
@@ -113,7 +109,7 @@ export const StepFormDialog: FC<Props> = ({
           closeDisabled={closeDisabled}
           submitLabel={submitLabel}
           onClickClose={actualOnClickClose}
-          onSubmit={actualOnSubmitAction}
+          onSubmit={onDelegateSubmit}
           onClickBack={actualOnClickBack}
           responseStatus={responseStatus}
           decorators={decorators}